Refactor : member도메인의 패키지 구조 수정
- member하위 패키지에 있던 common패키지를 같은 계층의 위치로 이동시킴
- controller파일에서 쓰지 않는 MemberApiController 패키지를 삭제함
