+++ src/main/java/toyProject/toyProject01/common/ToyProjectErrorCode.java
@@ -1,4 +1,4 @@
-package toyProject.toyProject01.member.common;
+package toyProject.toyProject01.common;
 
 import lombok.Getter;
