+++ src/main/java/toyProject/toyProject01/member/adapter/in/web/JoinApiController.java
@@ -11,8 +11,8 @@ import org.springframework.web.bind.annotation.RestController;
 import toyProject.toyProject01.member.application.port.in.MemberJoinUseCase;
 import toyProject.toyProject01.member.application.port.in.command.JoinCommand;
 import toyProject.toyProject01.member.application.service.MemberServiceException;
-import toyProject.toyProject01.member.common.ResultDto;
-import toyProject.toyProject01.member.common.ToyProjectErrorCode;
+import toyProject.toyProject01.common.ResultDto;
+import toyProject.toyProject01.common.ToyProjectErrorCode;
 
 @Slf4j
 @RestController
