+++ src/main/java/toyProject/toyProject01/member/adapter/in/web/MemberController.java
@@ -2,17 +2,9 @@ package toyProject.toyProject01.member.adapter.in.web;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-import toyProject.toyProject01.member.application.port.in.MemberJoinUseCase;
-import toyProject.toyProject01.member.application.port.in.command.JoinCommand;
-import toyProject.toyProject01.member.common.ResultDto;
 
 @Slf4j
 @Controller
